refactor: split route detail into view + dedicated update page
Move the route edit form (target + auth config) off the detail page into
a new route/edit page. route/manage now shows the route read-only with
Visit + Edit actions and keeps the Danger Zone delete; route/edit holds
the editable form and returns to the detail page after saving via
route.createV2.
