+++ tests/route.spec.js
@@ -36,8 +36,7 @@ test.describe('routes', () => {
 	test('clicking a row opens the route detail page', async ({ page }) => {
 		await setMocks({
 			'route.list': { ok: true, result: { items: [sampleRoute] } },
-			'route.get': { ok: true, result: sampleRoute },
-			'deployment.list': { ok: true, result: { items: [sampleDeployment] } }
+			'route.get': { ok: true, result: sampleRoute }
 		})
 
 		await page.goto('/route?project=test-project')
@@ -50,7 +49,56 @@ test.describe('routes', () => {
 test.describe('route detail', () => {
 	const detailUrl = '/route/manage?project=test-project&location=gke&domain=example.com&path=%2F'
 
-	test('renders the route and seeds its config', async ({ page }) => {
+	test('renders the route read-only with edit and delete actions', async ({ page }) => {
+		await setMocks({
+			'route.get': { ok: true, result: sampleRoute }
+		})
+
+		await page.goto(detailUrl)
+
+		await expect(page.getByRole('heading', { name: 'Route detail' })).toBeVisible()
+		await expect(page.locator('#view-location')).toHaveValue('gke')
+		await expect(page.locator('#view-domain')).toHaveValue('example.com')
+		await expect(page.locator('#view-path')).toHaveValue('/')
+		// deployment://web is split into a friendly type + destination.
+		await expect(page.locator('#view-type')).toHaveValue('Deployment')
+		await expect(page.locator('#view-target')).toHaveValue('web')
+		await expect(page.locator('#view-auth')).toHaveValue('None')
+		// Editing lives on a dedicated page; deleting stays here.
+		await expect(page.getByRole('link', { name: 'Edit' }))
+			.toHaveAttribute('href', '/route/edit?project=test-project&location=gke&domain=example.com&path=%2F')
+		await expect(page.getByRole('button', { name: 'Delete route' })).toBeVisible()
+		// Read-only: no editable target/auth controls on the detail page.
+		await expect(page.locator('#input-auth')).toHaveCount(0)
+	})
+
+	test('delete confirms then calls route.delete', async ({ page }) => {
+		await setMocks({
+			'route.get': { ok: true, result: sampleRoute },
+			'route.delete': { ok: true, result: {} }
+		})
+
+		await page.goto(detailUrl)
+
+		await page.getByRole('button', { name: 'Delete route' }).click()
+		await page.locator('.swal2-confirm').click()
+
+		await expect.poll(async () => {
+			const log = await getRequestLog()
+			return log.some((r) => r.path === '/route.delete')
+		}).toBeTruthy()
+
+		const req = (await getRequestLog()).find((r) => r.path === '/route.delete')
+		if (!req) throw new Error('expected a route.delete request')
+		expect(JSON.parse(req.body)).toMatchObject({ location: 'gke', domain: 'example.com', path: '/' })
+		await expect(page).toHaveURL(/\/route\?project=test-project/)
+	})
+})
+
+test.describe('route edit', () => {
+	const editUrl = '/route/edit?project=test-project&location=gke&domain=example.com&path=%2F'
+
+	test('seeds the form from the existing route', async ({ page }) => {
 		await setMocks({
 			'route.get': {
 				ok: true,
@@ -62,19 +110,16 @@ test.describe('route detail', () => {
 			}
 		})
 
-		await page.goto(detailUrl)
+		await page.goto(editUrl)
 
-		await expect(page.getByRole('heading', { name: 'Route detail' })).toBeVisible()
-		await expect(page.locator('#input-location')).toHaveValue('gke')
-		await expect(page.locator('#input-domain')).toHaveValue('example.com')
-		await expect(page.locator('#input-path')).toHaveValue('/')
+		await expect(page.getByRole('heading', { name: 'Edit route' })).toBeVisible()
 		// redirect:// prefix is stripped into the value field.
 		await expect(page.locator('#input-target_value')).toHaveValue('https://elsewhere.example.com')
 		// Basic-auth config seeds the auth fields.
 		await expect(page.locator('#input-basic_auth_user')).toHaveValue('admin')
 	})
 
-	test('saves forward-auth config via route.createV2', async ({ page }) => {
+	test('saves forward-auth config via route.createV2 and returns to the detail page', async ({ page }) => {
 		await setMocks({
 			'route.get': {
 				ok: true,
@@ -83,7 +128,7 @@ test.describe('route detail', () => {
 			'route.createV2': { ok: true, result: {} }
 		})
 
-		await page.goto(detailUrl)
+		await page.goto(editUrl)
 
 		// Switch protection to forward auth and fill it in (the trailing blank
 		// line proves empty header entries are dropped before the request).
@@ -113,30 +158,21 @@ test.describe('route detail', () => {
 			authResponseHeaders: ['X-Auth']
 		})
 		expect(body.config.basicAuth).toBeNull()
+
+		// On success the edit page returns to the route detail page.
+		await expect(page).toHaveURL(/\/route\/manage\?project=test-project&location=gke&domain=example\.com&path=%2F/)
 	})
 
-	test('delete confirms then calls route.delete', async ({ page }) => {
+	test('editing a deployment route lists deployments', async ({ page }) => {
 		await setMocks({
-			'route.get': {
-				ok: true,
-				result: { ...sampleRoute, target: 'redirect://https://elsewhere.example.com', config: {} }
-			},
-			'route.delete': { ok: true, result: {} }
+			'route.get': { ok: true, result: sampleRoute },
+			'deployment.list': { ok: true, result: { items: [sampleDeployment] } }
 		})
 
-		await page.goto(detailUrl)
+		await page.goto(editUrl)
 
-		await page.getByRole('button', { name: 'Delete route' }).click()
-		await page.locator('.swal2-confirm').click()
-
-		await expect.poll(async () => {
-			const log = await getRequestLog()
-			return log.some((r) => r.path === '/route.delete')
-		}).toBeTruthy()
-
-		const req = (await getRequestLog()).find((r) => r.path === '/route.delete')
-		if (!req) throw new Error('expected a route.delete request')
-		expect(JSON.parse(req.body)).toMatchObject({ location: 'gke', domain: 'example.com', path: '/' })
-		await expect(page).toHaveURL(/\/route\?project=test-project/)
+		// Deployment target → the deployment picker is populated for the location.
+		await page.locator('#input-target_deployment').click()
+		await expect(page.getByRole('option', { name: 'web' })).toBeVisible()
 	})
 })
